Rename cloud demo install dialog to promo dialog
