+++ client/features/home/workspace-setup/CreateWorkspaceDialog.tsx
@@ -17,7 +17,7 @@ import { validateWorkspaceFolderName } from '@/lib/workspace-name'
 import { WORKSPACE_TYPE_ORDER } from '@/lib/workspace-types'
 import type { WorkspaceType } from '@/lib/types'
 
-import { InstallMoiDialog } from './InstallMoiDialog'
+import { CloudDemoPromoDialog } from './CloudDemoPromoDialog'
 import { WorkspaceAgentStep } from './WorkspaceAgentStep'
 import { WorkspaceDialogContent } from './WorkspaceDialogContent'
 
@@ -45,9 +45,9 @@ export function CreateWorkspaceDialog({ trigger }: CreateWorkspaceDialogProps) {
   const isCreating = createMutation.isPending
 
   // Cloud demo: creating workspaces is off — the same trigger opens the
-  // install-moi dialog instead. After every hook so the hook order is stable.
+  // cloud-demo promo dialog instead. After every hook so the hook order is stable.
   if (appConfig.cloudDemo) {
-    return <InstallMoiDialog trigger={trigger} />
+    return <CloudDemoPromoDialog trigger={trigger} />
   }
 
   function finish(workspaceId: string) {
